Add basic support for a billing system, refactor app
This commit uses a SQLAlchemy backed ORM to store and create invoices
for projects, updating them daily, and creating new invoices every
month. This is all performed via a package called Flask APscheduler,
which is itself built off of APscheduler, and allows the user to buld
cron-like functionality into the application to run tasks at particular
schedules. This is nicer than having to set up a celery task queue at
least for now, but may need to be revisited.

Currently, invoices account for all computational work done in AWS
instances for workflows which ended in the last month, as well as the
storage costs for all workflow outputs which were calculated in previous
months. Still to be done is accounting for the storage cost of workflow
outputs which were computed within the same billing cycle, so invoices
at this time are slightly below their real billing cost.