Metadata export to FC using BDBag stored on S3 (resolves #292)
This commits introduces a change in uploading metadata to a FireCloud
workspace from Boardwalk by uploading the BDBag to an S3 storage
location and returning a presigned URL to the caller.

Specifically, we modified the route /repository/files/export (method
get_manifest). That endpoint originally created a tab-separated
values (TSV) file from a metadata manifest, and downloaded it
to a user's local system. That original functionality is preserved.

Here we added a query parameter to that endpoint so the user can specify
whether the original functionality should be executed, or whether
the metadata in form of a zipped BDBag should be uploaded to
Amazon AWS S3 for temporary storage. The two options can be invoked
with strings `tsv` and `bdbag`, respectively, as query parameters.

The commit also contains a class S3_FileHandler, which provides
the method used for S3 upload and generation of the presigned URL,
accompanied by a test suite. The compressed BDBag and the presigned URL
are created in the newly added method `bdbag_response`.

The newly added code required packages boto3, botocore, and mock. These
together with their dependecies have been added to `requirements.txt`.