Add auth decorator (resolves #291)
This auth decorator passes the session and any
authorization header on to the authorization
request check to dashboard. (also some headers
that ensure the request will make it through
session protection.

If dashboard give the OK, then the decorated
calls are allowed to run.

Add some unit tests for auth header

Unfortunately these tests are not hooked up into the current
testing framework, but I ran locally and they pass.