Redirect clients to download content from S3 (fixes #356)
In order to resolve the issue around the limit on the size of the response payload for AWS Lambda (6 MB at the time of implementation) and ensure that the response is achievable within 30 seconds, this updates makes the manifest API utilize a private S3 bucket as an intermediate storage for downloadable content and redirect the client to that downloadable content on S3 via an associated pre-signed URL with HTTP 302.

The tests use Minio as S3 replacement in test in the initial stage of development and then use `moto` in the final version.

Replaced google-compute-engine with BOTO_CONFIG

Based on recommendation from @hannes-ucsc, to fix the broken test on Travis CI, BOTO_CONFIG will be set to /dev/null so that we don't have to install google-compute-engine. See https://github.com/travis-ci/travis-ci/issues/7940#issuecomment-311411559 for more details.