Obtain program name and project name from top-level metadata
The `program` value was previously hardcoded to `TOPMed` and now we
are loading data from other programs, so the `program` value needs to be
obtained from the metadata.
The `project` value was previously obtained from nested metadata values,
and now it is obtained from a top-level value. This will provide more convenient
control of the `project` value going forward.