Address token deficit and surplus (fixes #390)
Deficits are handled by adding a `nudge` lambda that periodically supplies more tokens if needed. The number of missing tokens is an estimate. Supplying an estimated amount of tokens could result in a surplus which is addressed by expiring tokens similar to how SQS expired messages. We don't use SQS' expiration because that would require raising an exception in the `write` lambda. The backtraces from those exceptions would pollute the logs and the expired tokens would clog up the fail queues, drowning out messages returned from invocations of the `write` lambda that truly failed.