Handle more than one cell suspension per specimen (fixes #468)
This also changes how we compose the specimen document. Instead of doing it in two passes, with generic merging in the second pass, there now is one pass with merging strategies specific to each field. This reuses the accumulator mechanism we added for document aggregation. The generic merging broke in the multiple cell suspension case. This also fixes a giant bug with diamonds in a biomaterial DAG. For example, the old code would merge the same cell suspension more than once if it was reachable from the specimen via two downward edges. Oops.