Add export format description to README and update Boardwalk dev URL
The Boardwalk webservice UI has two "export" buttons, one to
download the metadata manifest to the local system as TSV,
and a second, labeled "Export to FireCloud", which uploads
a metadata manifest to Broad's FireCloud platform.

In the backend this is controlled by a new format parameter
which has two values, "tsv" and "bdbag".

The following was not part of the original story, but came up
while I worked on it:

In addition to that I replaced the URL "http://ucsc-cgp.org",
which is not running the CGP platform anymore, by
"https://commons.ucsc-cgp-dev.org" in the entire Webservice
Endpoints paragraph.