Refactor plugin mechanism (fixes #420)
Eliminate BaseIndexerProperties and introduce a Plugin class as the sole
entry point for all functionality customizable by a plugin. This design
is more extensible and eliminates always having to load and instantiate
two classes (the indexer and the indexer properties). At some point we
will add the ability to customize service functionality this way, too.

Document plugin mechanism.