Changed Java install in Dockerfile
Running `docker build -t ...` leads to an error in the installation
of Java. I think this was related to an obsolete GPG key
(EEA14886), which in turn is probably related to a new built
of the oracle-java8-installer from the webupd8team PPA, on
2018-10-17.

The new key, C2518248EEA14886, still has part of the old key.

But only changing the GPG key was insufficient to finish the
installation. I also needed to change the syntax to export the
key (details can be found here:
https://ram.tianon.xyz/post/2016/12/07/docker-setup.html)

Further, the SSL certificate update now follows the installation
of curl, dnsutils and Java as opposed to installation of all
packages in one command (apt update precedes the installation
of ca-certificates).