Add deletion to integration test (#838)
The fix works by changing the test reindexing to also assign a new
uuid for the test bundles. These new uuids are then searched in the
index to make sure that indexing was successful. Finally deletions for
all of the relevant bundles that were found in the index are sent to
Azul