Support duplicate notification (#947)
Duplicate notifications are supported by changing how contributions are
written to the index. First contributions are written with 'create'
which will fail if these contributions already exist. They will be
logged and then indexing tried again with update. This mean that the
tally counts will now be accurate except in a few rare cases.

If a notification has been indexed and aggregated and then a redundant
notification is indexed, the tally can have a count of 0 (since we only
update the tally count original contributions). For this reason I also
added a test that ensures that aggregation happens if the tally is 0.