Fix: OpenAPI UI renders incorrect server URL (#1262)
Even though API Gateway is configured with multiple (namely 2) different
server domains, when requesting the OpenAPI spec from AWS only 1 is
returned. My guess is that because OpenAPI v2 only supported one "host"
entry, AWS doesn't preserve this information for when the spec is
generated, even if v3 is requested. Probably, behind the scenes, they
just convert their v2 spec to v3.

To solve this, I hard coded the URL in the OpenAPI template. and removed
from the spec any servers that are API Gateway lists.