Explicitly document path and method separately (#1436, #1454)
The bulk of this work is directed towards solving #1436, but #1454 is
also incidentally solved.

Now the specifications for `path_spec` and `method_spec` are separate.
This reduces duplication in the final OpenAPI spec. Unfortunately some
duplication is still present in the example health endpoint.

According to the OpenAPI Specification 3.0,
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.0.md#path-item-object
a Path Item Object is limited in its fields and cannot contain all
fields that are shared between HTTP methods.