Update Python to 3.8.2
In Python 3.7, building of .pyc files is reproducible. In other words,
for the same .py file, you will always get the same .pyc file. This is
necessary because .pyc files in the lambda's vendor directory are
included by Chalice in the deployment package.

If we're updating Python, we figured we might was well go all the way.