TDR client should not check for table at initialization
A check like this limits the versatility of the client. If a table is missing we will learn soon enough. The true reason the check existed seemed to have been to force eager initialization of a cached property so that it coincided with context manager being active. I consider that an anti-pattern.