Fix: Manifest integration test ignores redirect responses
In order to handle Retry-After for redirects we initially patched the urllib3 retry policy for the `requests` library to 1) perform retries on 301 status codes and 2) honor the Retry-After header. This caused urllib3 to treat 301 as an error and to retry. At that time, this was the only way to handle redirects with Retry-After. Since then, we upgraded urllib3 to a version that refined the handling of redirects to handle Retry-After natively. The retry on error logic was also changed to disregard the Location header. The latter change caused the IT to retry the original manifest URL instead of following the redirect. Because the service looks up cached manifests before kicking off a manifest generation, these "retries" eventually resulted in a 302 to the manifest file, once the generation from the original requests completed.