Fix:  `previous` link broken with page size of one (#3732)
With a page size of one, the `search_before` and `search_after` variables in `_generate_paging_dict()` point at the same value, causing `search_before` to be serialized as JSON twice, from "…" to '"…"' to '"\\"…\\"'. That extra backslash causes the ES query to return no results which leads to the OOB in the ticket.

Test coverage of this case will be added as part of the solution to #3715.