Fix: `make requirements_update` is not idempotent
If the first invocation dirties the working copy by modifying the transitive requirement files, the second invocation will not use the Docker images cached from the first. This is slow and, if enough time passes between the first and second invocation, may lead to more changes to the transitive requirement files. The latter is OK but I'd rather have the behavior be fast and predictable. To force an update, one can always use the `requirements_update_force` target.