Rename `lambda` to `app` in api_gateway.tf.json.template.py
This is the first step in cleaning up the conflation between application and lambda. Until now, we imprecisely used the term *lambda* to be synonymous with AWS Lambda function. But what we really mean is an entire application that is composed of multiple such functions.

Ultimately this will lead to the `lambdas` directory being renamed to `apps`.