[p] Remove incomplete authentication for cached manifests (#3328, #4050)
Due to the uncertainty of how to properly implement that feature, it's not worth dragging dead code and some partial implementation around.

This may remove the need for consolidating the parsing of the object key since it removes one of the places were a manifest's object key is parsed. I just don't know how many places are left.