[R] Ran `make requirements_update`
This branch does not modify anything that would require doing this but, for some reason the corrsponding check fails on GitLab `anvildev` so I'm doing this to get past that. The fact that the check only fails on anvildev is indicative that the problem is with anvildev, not this branch.

https://gitlab.anvil.gi.ucsc.edu/ucsc/azul/-/jobs/29219
(cherry picked from commit 7ac6a2c9f503ca356fb9fa6cee397d1456bde691)