Fix mismatched entity IDs in canned TDR table
Some canned table rows contain conflicting values for the entities' UUIDs: the
primary keys do not match the document IDs in the `content` column. We suspect
this is the result of a bug in the recently-retired `recan_bundle_tdr` script.
Soon, we will replace file names in HCA bundles with entity references, at
which point these inconsistent UUIDs would break the build.